requirements.txt
django==3.0.6
djangorestframework==3.11.0
celery==4.4.2
redis==3.5.0
requests